--- src/Onyx.Engine/Application.cpp
@@ -1,11 +1,0 @@
-#include "Application.h"
-
-Application::Application()
-{
-  std::cout << "Onyx::Application";
-}
-
-Application::~Application()
-{
-
-}
