--- src/Onyx.Engine/Application.h
@@ -1,9 +1,0 @@
-#pragma once
-#include <iostream>
-
-class Application
-{
-public:
-  Application();
-  ~Application();
-};
