+++ src/Onyx.Engine/System/Application.cpp
@@ -1,0 +1,14 @@
+#include "Application.h"
+
+namespace Onyx::Engine::System
+{
+  Application::Application()
+  {
+    std::cout << "Onyx::Application";
+  }
+  
+  Application::~Application()
+  {
+  
+  }
+}
