+++ src/Onyx.Engine/System/Application.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <iostream>
+
+namespace Onyx::Engine::System
+{
+  class Application
+  {
+  public:
+    Application();
+    ~Application();
+  };
+}
